Started with the modularization of components in React.
